Improve container startup
The backingFilePath attribute of MetadataProvider did not refer to the now
removed file and thus did not help in providing metadata on first startup
of the SP, if the IdP could not be reached, e.g. because it was still
starting up. Instead, give a short minRefreshDelay to poll the metadata
more quickly as soon as the IdP is ready.