Use PostgreSQL database with Keycloak
Since Keycloak uses liquibase for database schema management, we have to keep
the database privileges lax for the schema, because liquibase needs to be able
to create tables just for not making the server startup fail, even if we
would configure manual initialization and updates of the schema (!).

There can still be timing problems when starting with docker-compose because
Keycloak just exits if it does not immediately get a database connection,
e.g. because the database is still starting up. A more healthy behavior would
be to retry to connect to the database.

Configuration of realm, client and example user is now done during startup
of Keycloak, because it requires the database to be available, which is no
more the case during image build.