Reduce code duplication for checking changes
Also remove the differentiation between processType 'edit' and
'copy' because the 'copy' part was only used for the save button.
There it is better to let the user receive an error from the
server which explains that the user cannot create a new user with
identical username and/or email address than to just disable the
button to save the new user.