Fix field for coordinates
The text in the select field for the coordinates wasn't reset
when opening form to create a new institution or to edit an
existing one. Empty the list of coordinates in the coordinates
store wasn't enough because the list of the select became empty
but the text remained. Therefore, introduce a new ref.