Pass resetNotification to resetForm function
Before, we imported the function resetNotification() directly in
use-form.js but this also meant that we worked with a different copy
from it. The resetNotification() in use-form.js was changing a
different hasRequestError variable than the one we changed by using
resetNotification() in the form components like ManageInstitution.vue
where we also imported an(other) object from use-notification.js.