Display errors of bean validation in the client
What we did with the validation errors coming from keycloak for the
profile attributes we now extend to other dialogs as well. Because of
this many methods are now placed in use-form.js which make them
easy to change. This commit also renames some of the existing methods
and variables to make it more clear what they are used for.