Make ChipTextField update attribute of institution
Since we don't save the input of the Textfield directly
we have to watch the changes made to the array which is
visualized by the chips.

Additionally, use a "real" button to add new entries because buttons
offer better usability. With them users can see directly when they are
allowed to press them.