Show username field only once at the top
Before, we had a non-generic textfield for the username
because we wanted it to be at the top of the form because it doesn't
fit into the categories name or contact but because it's an
identifier it shouldn't be placed under "miscellaneous". In commit
a864edccbc5e7f15552d7cf2030cedec6f88b728 the username
was added as an attribute to the user_profile.json. In this case
it would be handled the generic way and appear a second time under
"miscellaneous" which we prevent by this commit.