Always load users/institutions
Before, we didn't load them e.g. when creating new users
because we got the new user from the POST request and it was
not necessary to make another request. We already had all data
we needed.
Now, we offer functions to sort and filter tables so new objects
also have to be sorted/filtered. Since the backend does this
already we just send another request to it.