Moving focus to next element when using tab
Without this measurement when a user would press tab the focus
would get lost so if the user presses tab again the first element
of the form would be focused, not the next (expected) one.

Since we now have to listen to the events "updated:focused" and
"keydown.tab" we use the variable "isAdding" to prevent that the
current input is added twice.