Improve usability of filters for user attributes
Problem is that the filtering of user attributes which are **arrays**
works only with an exact value. That's why this commit chooses select
components in these cases.

Columns like 'active' theoretically don't need a filter since
ordering the table would also show the same result. But to keep the
application consistent booleans can also be filtered via select
components.