Fix filter by boolean attributes
With Hibernate 6.6 it is no longer possible to use `as` to cast on
the database side. With Hibernate 7 the `cast` function can be
directly used on the expression. For now this inelegant cast
must be used. For more information, see the following issue and the
migration guide.

Issue-Report: https://hibernate.atlassian.net/browse/HHH-18710
Migration-Guide: https://docs.jboss.org/hibernate/orm/6.6/migration-guide/migration-guide.html#criteria-query