Fix error because of copied user without title
The user attributes have to be initialized with `undefined`
because Keycloak allows only strings inside of arrays so we cannot
initialize them with an empty string.

It's also no good idea to initialize them with an empty string inside
of an array because in this case the form fields look like there was
already entered something. The placeholder is not inside the field
but at the top like in other fields with some value.