Automatically apply rule if field is required
To reduce code the components have to find out the translation
for the error message on their own. To enable them to do so we
use provide/inject and pass the top-level property of the locale
object. This way we don't have to pass one prop per component
but can pass the property once per dialog. To prevent future
intersections the labels for the export dialog get their own
property.