Fix selection of institutions for users
Before, only a number of items (users/institutions) was loaded
that equals the number in "itemsPerPage". Then it was only to
select from this number of institutions in ManageUser.

Now we ensure that all users/institutions are loaded at the
beginning and if the user filters the results or changes the
"itemsPerPage" etc. it will still be possible to select any of
the available institutions in the backend.