Change way that textfield is re-focused
Before, it was necessary that the ChipTextField is inside a form
to be able to focus the previous element but for unit tests of the
ChipTextField we don't need a whole form.

The new mechanism is also less error-prone because it allows
non-input elements to be "between" input and button without breaking
the re-focus.