Add test to check translation of role in table
Put roles in sharedTests.js to prevent code duplication.

Add columns as new prop in UserTable. I didn't find a better solution
to be able to overwrite the columns from the test as it is not
possible to mock simple functions. The functions always have to be part of an
object or composable to be able to mock them.