Fix bug with v-autocomplete fields in table header
If you focused an autocomplete field and pressed the arrow-down-key
the dropdown of that field would be opened but the focus would
switch to the next field.

It seems that components inside v-lists automatically move the focus
to the next field when the arrow-down-key is pressed. This behavior
interfered with the the autocomplete that opened the dropdown when
that key was pressed.